+++ Sources/OpenKey/macOS/ModernKey/ConvertToolViewController.h
@@ -1,36 +1,37 @@
 //
 //  ConvertToolViewController.h
-//  OpenKey
+//  MyOpenKey
 //
 //  Created by Tuyen on 9/4/19.
 //  Copyright © 2019 Tuyen Mai. All rights reserved.
+//  Modified 2026 by Huỳnh Quốc Đạt for MyOpenKey (GPLv3).
 //
 
 #import <Cocoa/Cocoa.h>
-#import "MyTextField.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface ConvertToolViewController : NSViewController<MyTextFieldDelegate>
-@property (weak) IBOutlet NSButton *AlertWhenComplete;
-@property (weak) IBOutlet NSButton *ToAllCaps;
-@property (weak) IBOutlet NSButton *ToNonCaps;
-@property (weak) IBOutlet NSButton *ToCapsFirstLetter;
-@property (weak) IBOutlet NSButton *ToCapsCharEachWord;
-@property (weak) IBOutlet NSButton *ToRemoveSign;
+@interface ConvertToolBridge : NSObject
 
-@property (weak) IBOutlet NSButton *ConvertInClipBoard;
++ (instancetype)shared;
 
-@property (weak) IBOutlet NSButton *SControl;
-@property (weak) IBOutlet NSButton *SOption;
-@property (weak) IBOutlet NSButton *SCommand;
-@property (weak) IBOutlet NSButton *SShift;
-@property (weak) IBOutlet MyTextField *SHotKey;
+@property (nonatomic, assign) NSInteger fromCode;
+@property (nonatomic, assign) NSInteger toCode;
+@property (nonatomic, readonly) NSArray<NSString *> *availableCodeTables;
 
-@property (weak) IBOutlet NSPopUpButton *FromCode;
-@property (weak) IBOutlet NSPopUpButton *ToCode;
-@property (weak) IBOutlet NSButton *ReverseCode;
+@property (nonatomic, assign) NSInteger caseOption; // 0: Giữ nguyên, 1: SANG CHỮ HOA, 2: sang chữ thường, 3: Viết hoa chữ cái đầu câu, 4: Viết Hoa Mỗi Từ
+@property (nonatomic, assign) BOOL removeMark;
+@property (nonatomic, assign) BOOL alertWhenCompleted;
 
+@property (nonatomic, assign) NSInteger hotKeyPreset;
+@property (nonatomic, readonly) NSArray<NSString *> *availableHotKeyPresets;
+
+- (void)reverseCodes;
+- (BOOL)convertClipboard:(nullable NSWindow *)window;
+
+@end
+
+@interface ConvertToolViewController : NSViewController
 @end
 
 NS_ASSUME_NONNULL_END