+++ Sources/OpenKey/macOS/ModernKey/ModernKey-Bridging-Header.h
@@ -5,3 +5,4 @@
 
 #import "OpenKeyBridge.h"
 #import "MacroViewController.h"
+#import "ConvertToolViewController.h"